check availability of localStorage before use

+++ src/store.js
@@ -1,6 +1,6 @@
 import Vue from 'vue';
 import Vuex from 'vuex';
-import { getJsonStorage, vehicles } from '@/lib/utils';
+import { getJsonStorage, vehicles, localStorageAvailable } from '@/lib/utils';
 import { fetchRouteChanges } from '@/lib/fetch';
 
 Vue.use(Vuex);
@@ -24,6 +24,8 @@ export default new Vuex.Store({
         routeChanges: {},
         routeChangesLoading: false,
         routeChangesFetched: null,
+        // localStorage might not be available eg. android browser runs in private mode
+        localStorageAvailable: localStorageAvailable(),
     },
     getters: {
         chosenMots(state) {
@@ -55,28 +57,32 @@ export default new Vuex.Store({
             const idx = state.favoriteStations.findIndex(s => s.id === station.id);
             if (idx > -1) return;
             state.favoriteStations.push(station);
+            if (!this.localStorageAvailable) return;
             localStorage.setItem('favoriteStations', JSON.stringify(state.favoriteStations));
         },
         removeStation(state, station) {
             const idx = state.favoriteStations.findIndex(s => s.id === station.id);
             if (idx === -1) return;
             state.favoriteStations.splice(idx, 1);
+            if (!this.localStorageAvailable) return;
             localStorage.setItem('favoriteStations', JSON.stringify(state.favoriteStations));
         },
         addVehicle(state, vehicle) {
             const idx = state.chosenVehicles.indexOf(vehicle);
             if (idx > -1) return;
             state.chosenVehicles.push(vehicle);
+            if (!this.localStorageAvailable) return;
             localStorage.setItem('chosenVehicles', JSON.stringify(state.chosenVehicles));
         },
         removeVehicle(state, vehicle) {
             const idx = state.chosenVehicles.indexOf(vehicle);
             if (idx === -1) return;
             state.chosenVehicles.splice(idx, 1);
+            if (!this.localStorageAvailable) return;
             localStorage.setItem('chosenVehicles', JSON.stringify(state.chosenVehicles));
         },
         clearStorage(state) {
-            localStorage.clear();
+            if (this.localStorageAvailable) localStorage.clear();
             state.favoriteStations = defaultStates.favoriteStations();
             state.chosenVehicles = defaultStates.chosenVehicles();
         },
